Emit per-policy logs instead of summary logs
This produces one log per policy, for each incoming event.  The `violation`
field will be true if the resouce violated the policy.  Similarly,
`remediation` will be true if the remediation was successful.

This also ensure that each policy appears in the logs, so audit tools can
depend on having evidence from every implemented policy.